+++ R/example.R
@@ -1,12 +1,14 @@
 #' NBN demonstration dataset
 #'
 #' @description
-#' This dataset contains 100 bird records recorded within three National Nature
-#' Reserves in 2008, for the purpose of demonstrating the functionality of the
-#' NBN Gateway. This website collated and made freely available biological species records
-#' across the UK. It was superseded in March 2017 by the [NBN Atlas](https://nbnatlas.org/)
+#' This dataset contains sightings of 10 common birds species within three national
+#' nature reserves, Monks Wood, Woodwalton Fen and Holme Fen, occuring in or adjacent
+#' to TL28 10km grid square, recorded in 2008 for the purpose of demonstrating the
+#' functionality of the NBN Gateway. This website collated and made freely available
+#' biological species records across the UK. It was superseded in March 2017 by
+#' the [NBN Atlas](https://nbnatlas.org/)
 #'
-#' @format A tibble with 100 rows and 16 variables
+#' @format A tibble with 100 rows and 17 variables
 #' \describe{
 #'   \item{RecordKey}{primary key associated with the occurrence record}
 #'   \item{SurveyKey}{identifier key for the survey}
@@ -22,7 +24,8 @@
 #'   \item{SiteName}{name of the site or location where the species was recorded}
 #'   \item{Recorder}{name or list of names for one or more recorders for the species record}
 #'   \item{Determiner}{name or list of names for one or more determiners for the species record}
-#'   \item{Abundance}{attribute field of number of individuals recorded}
-#'   \item{SpeciesName}{attribute field of scientific name of species recorded}
+#'   \item{Abundance}{attribute field containing number of individuals recorded}
+#'   \item{TaxonName}{attribute field containing scientific name of species recorded}
+#'   \item{CommonName}{attribute field containing common name of species recorded}
 #' }
 "nbn_demonstration_dataset"